requirements.txt
opencv-python~=4.10.0.84
numpy~=1.26.4
Flask~=3.0.3
ultralytics~=8.2.93
pip~=24.0
pillow~=10.4.0
Jinja2~=3.1.4
setuptools~=65.5.0
packaging~=24.1
pyparsing~=3.1.4
typing_extensions~=4.12.2
matplotlib~=3.9.2
requests~=2.32.3
beautifulsoup4~=4.12.3
pycocotools~=2.0.8
scikit-image~=0.24.0
tqdm~=4.66.5
filelock~=3.16.0
pytz~=2024.2
torch~=2.4.1
ultralytics-thop~=2.0.6
colorama~=0.4.6
PyYAML~=6.0.2
charset-normalizer~=3.3.2
scipy~=1.14.1
mpmath~=1.3.0
sympy~=1.13.2
torchvision~=0.19.1
psutil~=6.0.0
urllib3~=2.2.3
fsspec~=2024.9.0
MarkupSafe~=2.1.5
pandas~=2.2.2
python-dateutil~=2.9.0.post0
certifi~=2024.8.30
py-cpuinfo~=9.0.0
seaborn~=0.13.2
cycler~=0.12.1
six~=1.16.0
networkx~=3.3
idna~=3.8
contourpy~=1.3.0
fonttools~=4.53.1
kiwisolver~=1.4.7